tests/submission/Infraware/Forms/contents/Forms/input_name.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_name</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='hidden' id='input_text' name='name'></p>
	</form>

	<script>
		
		var input_text = document.getElementById("input_text");
		name = input_text.name;
		input_text.name = "input_text";

		if (typeof(input_text.name) == "string") {
			if (name == 'name' && input_text.name == "input_text") {
				test(function() {
					assert_true(true, "name attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("name attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("name attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
